// src/shared/constants/enums.ts
export enum Gender {
  Male,
  Female,
}
// ...
export enum ApplyForAcademicYear {
  Year2024,
  Year2025,
  Year2026,
}

export const applyForAcademicYearOptions = Object.keys(ApplyForAcademicYear)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key.replace("Year", ""),
    value: value.toString(),
  }));

export enum IntakeTerm {
  Semester1,
  Semester2,
  Semester3,
}

export const applyForIntakeTermsOptions = Object.keys(IntakeTerm)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key.replace("Semester", "Semester "),
    value: value.toString(),
  }));

export enum CurrentLivingStatus {
  Parents,
  Father,
  Mother,
  Guardian,
}

export const applyForCurrentLivingStatusOptions = Object.keys(CurrentLivingStatus)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key,
    value: value.toString(),
  }));

export enum BloodType {
  A,
  B,
  AB,
  O,
}

export const applyForBloodTypeOptions = Object.keys(BloodType)
  .filter((key) => isNaN(Number(key)))
  .map((key, value) => ({
    label: key,
    value: value.toString(),
  }));

export enum IncomingGrade {
  Nursery1 = "Nursery 1",
  Nursery2 = "Nursery 2",
  K1 = "K1",
  K2 = "K2",
  Grade1 = "Grade 1",
  Grade2 = "Grade 2",
  Grade3 = "Grade 3",
  Grade4 = "Grade 4",
  Grade5 = "Grade 5",
  Grade6 = "Grade 6",
  Grade7 = "Grade 7",
  Grade8 = "Grade 8",
  Grade9 = "Grade 9",
  Grade10 = "Grade 10",
  Grade11 = "Grade 11",
  Grade12 = "Grade 12",
}
// ...
export enum StudentSpokenLanguage {
  Bengali = "Bengali",
  Bhojpuri = "Bhojpuri",
  Burmese = "Burmese",
  EgyptianArabic = "Egyptian Arabic",
  English = "English",
  French = "French",
  Gujarati = "Gujarati",
  Hausa = "Hausa",
  Hindi = "Hindi",
  IranianPersian = "Iranian Persian",
  Italian = "Italian",
  Japanese = "Japanese",
  Javanese = "Javanese",
  Korean = "Korean",
  MandarinChinese = "Mandarin Chinese",
  Marathi = "Marathi",
  Portuguese = "Portuguese",
  Russian = "Russian",
  Spanish = "Spanish",
  StandardGerman = "Standard German",
  Tamil = "Tamil",
  Telugu = "Telugu",
  Thai = "Thai",
  Turkish = "Turkish",
  Urdu = "Urdu",
  Vietnamese = "Vietnamese",
  WesternPunjabi = "Western Punjabi",
  WuChinese = "Wu Chinese",
  YueChinese = "Yue Chinese",
}

export enum PreferredLanguage {
  Thai = "Thai",
  Japanese = "Japanese",
  Korean = "Korean",
  Chinese = "Chinese",
}

export enum StudentEnglishProficiency {
  Native = "Native",
  Fluent = "Fluent",
  Good = "Good",
  Fair = "Fair",
  None = "None",
}